reporter-web-app: Use Rescripts instead ejecting Create React App
Replace current 'ejected' (fork) of Create React App[1] with the upstream
version of which its config has been extended using Rescripts[2]. This
makes updating to newer versions of Create React App easier.

[1]: https://github.com/facebook/create-react-app
[2]: https://github.com/harrysolovay/rescripts
