--- reporter-web-app/.eslintrc.json
@@ -1,71 +1,0 @@
-{
-    "parser": "babel-eslint",
-    "extends": ["eslint-config-airbnb"],
-    "env": {
-        "browser": true,
-        "es6": true
-    },
-    "parserOptions": {
-        "ecmaFeatures": {
-            "jsx": true
-        },
-        "ecmaVersion": 2018,
-        "sourceType": "module"
-    },
-    "plugins": [
-        "react",
-        "babel"
-    ],
-    "rules": {
-        "camelcase": [
-            "error", {
-              "ignoreDestructuring": true,
-              "allow": [ "^UNSAFE_" ]
-            }
-        ],
-        "comma-dangle": ["error", {
-            "arrays": "never",
-            "objects": "never",
-            "imports": "never",
-            "exports": "never",
-            "functions": "ignore"
-        }],
-        "indent": [
-            "error",
-            4
-        ],
-        "linebreak-style": [
-            "error",
-            "unix"
-        ],
-        "max-len": ["error", { "code": 120 }],
-        "no-plusplus": [2, { "allowForLoopAfterthoughts": true }],
-        "no-underscore-dangle": ["error", {
-            "allow": [
-                "_id"
-            ]
-        }],
-        "semi": [
-            "error",
-            "always"
-        ],
-        "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }],
-        "react/jsx-indent": [2, 4],
-        "react/jsx-indent-props": [2, 4],
-        "react/sort-comp": [2, {
-            "order": [
-              "static-methods",
-              "instance-variables",
-              "lifecycle",
-              "/^on.+$/",
-              "everything-else",
-              "render"
-            ]
-          }],
-        "react/forbid-prop-types": [0, {}],
-        "quotes": [
-            "error",
-            "single"
-        ]
-    }
-}
