+++ reporter-web-app/package.json
@@ -13,6 +13,12 @@
         "url": "https://github.com/oss-review-toolkit/ort.git",
         "directory": "reporter-web-app"
     },
+    "scripts": {
+        "start": "rescripts start",
+        "build": "rescripts build",
+        "test": "rescripts test",
+        "lint": "eslint ./src"
+    },
     "dependencies": {
         "@ant-design/icons": "^4.5.0",
         "antd": "^4.14.0",
@@ -21,10 +27,11 @@
         "pako": "^2.0.3",
         "postcss": "^8.2.8",
         "prop-types": "^15.7.2",
-        "react": "^17.0.1",
+        "react": "^17.0.2",
         "react-app-polyfill": "^2.0.0",
-        "react-dom": "^17.0.1",
+        "react-dom": "^17.0.2",
         "react-redux": "^7.2.2",
+        "react-scripts": "4.0.3",
         "react-syntax-highlighter": "^15.4.3",
         "recharts": "2.0.8",
         "redux": "^4.0.5",
@@ -35,137 +42,32 @@
         "spdx-correct": "^3.1.1",
         "spdx-license-list": "^6.4.0"
     },
-    "scripts": {
-        "start": "node scripts/start.js",
-        "build": "node scripts/build.js",
-        "test": "node scripts/test.js",
-        "lint": "eslint ./src"
-    },
     "devDependencies": {
-        "@babel/core": "7.13.10",
-        "@svgr/webpack": "5.5.0",
-        "@testing-library/jest-dom": "^5.11.9",
-        "@testing-library/react": "^11.2.5",
-        "@testing-library/user-event": "^13.0.1",
-        "@typescript-eslint/eslint-plugin": "^4.18.0",
-        "@typescript-eslint/parser": "^4.18.0",
-        "babel-eslint": "10.1.0",
-        "babel-jest": "26.6.3",
-        "babel-loader": "8.2.2",
-        "babel-plugin-named-asset-import": "^0.3.7",
-        "babel-preset-react-app": "^10.0.0",
-        "case-sensitive-paths-webpack-plugin": "2.4.0",
-        "css-loader": "5.1.3",
-        "dotenv": "8.2.0",
-        "dotenv-expand": "5.1.0",
-        "eslint": "7.22.0",
-        "eslint-config-airbnb": "18.2.1",
-        "eslint-config-react-app": "^6.0.0",
-        "eslint-loader": "4.0.2",
-        "eslint-plugin-babel": "5.3.1",
-        "eslint-plugin-flowtype": "5.4.0",
-        "eslint-plugin-import": "2.22.1",
-        "eslint-plugin-jsx-a11y": "6.4.1",
-        "eslint-plugin-react": "7.22.0",
-        "eslint-plugin-react-hooks": "4.2.0",
-        "file-loader": "6.2.0",
-        "fs-extra": "9.1.0",
-        "html-webpack-inline-source-plugin": "1.0.0-beta.2",
-        "html-webpack-plugin": "4.5.2",
-        "identity-obj-proxy": "3.0.0",
-        "jest": "26.6.3",
-        "jest-environment-jsdom-fourteen": "1.0.1",
-        "jest-resolve": "26.6.2",
-        "jest-watch-typeahead": "^0.6.1",
-        "mini-css-extract-plugin": "1.3.9",
-        "optimize-css-assets-webpack-plugin": "5.0.4",
-        "pnp-webpack-plugin": "^1.6.4",
-        "postcss-flexbugs-fixes": "5.0.2",
-        "postcss-loader": "4.2.0",
-        "postcss-normalize": "^9.0.0",
-        "postcss-preset-env": "6.7.0",
-        "postcss-safe-parser": "5.0.2",
-        "react-dev-utils": "^11.0.4",
-        "resolve-url-loader": "^3.1.2",
-        "sass-loader": "11.0.1",
-        "style-loader": "2.0.0",
-        "terser-webpack-plugin": "4.2.3",
-        "typescript": "^4.2.3",
-        "url-loader": "4.1.1",
-        "webpack": "4.44.2",
-        "webpack-dev-server": "3.11.2",
-        "webpack-event-plugin": "^1.1.1",
-        "webpack-manifest-plugin": "3.1.0"
+        "@testing-library/jest-dom": "^5.11.4",
+        "@testing-library/react": "^11.1.0",
+        "@testing-library/user-event": "^12.1.10",
+        "@rescripts/cli": "^0.0.16",
+        "fs-extra": "^9.1.0",
+        "html-inline-css-webpack-plugin": "^1.11.0",
+        "webpack-event-plugin": "^1.1.1"
     },
-    "babel": {
-        "presets": [
-            "react-app"
+    "eslintConfig": {
+        "extends": [
+            "react-app",
+            "react-app/jest"
         ]
     },
     "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
+    "production": [
+        ">0.2%",
+        "not dead",
+        "not op_mini all"
+    ],
+    "development": [
+        "last 1 chrome version",
+        "last 1 firefox version",
+        "last 1 safari version"
+    ]
     },
-    "eslintConfig": {
-        "extends": "react-app"
-    },
-    "jest": {
-        "roots": [
-            "<rootDir>/src"
-        ],
-        "collectCoverageFrom": [
-            "src/**/*.{js,jsx,ts,tsx}",
-            "!src/**/*.d.ts"
-        ],
-        "setupFiles": [
-            "react-app-polyfill/jsdom"
-        ],
-        "setupFilesAfterEnv": [
-            "<rootDir>/src/setupTests.js"
-        ],
-        "testMatch": [
-            "<rootDir>/src/**/__tests__/**/*.{js,jsx,ts,tsx}",
-            "<rootDir>/src/**/*.{spec,test}.{js,jsx,ts,tsx}"
-        ],
-        "testEnvironment": "jest-environment-jsdom-fourteen",
-        "transform": {
-            "^.+\\.(js|jsx|ts|tsx)$": "<rootDir>/node_modules/babel-jest",
-            "^.+\\.css$": "<rootDir>/config/jest/cssTransform.js",
-            "^(?!.*\\.(js|jsx|ts|tsx|css|json)$)": "<rootDir>/config/jest/fileTransform.js"
-        },
-        "transformIgnorePatterns": [
-            "[/\\\\]node_modules[/\\\\].+\\.(js|jsx|ts|tsx)$",
-            "^.+\\.module\\.(css|sass|scss)$"
-        ],
-        "modulePaths": [],
-        "moduleNameMapper": {
-            "^react-native$": "react-native-web",
-            "^.+\\.module\\.(css|sass|scss)$": "identity-obj-proxy"
-        },
-        "moduleFileExtensions": [
-            "web.js",
-            "js",
-            "web.ts",
-            "ts",
-            "web.tsx",
-            "tsx",
-            "json",
-            "web.jsx",
-            "jsx",
-            "node"
-        ],
-        "watchPlugins": [
-            "jest-watch-typeahead/filename",
-            "jest-watch-typeahead/testname"
-        ]
-    },
-    "homepage": "."
+    "homepage": "./"
 }
